ai-service/requirements.txt
# AI Service Dependencies

fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pdfplumber==0.10.3
python-docx==1.1.0
openai==1.12.0
pydantic==2.6.1
python-dotenv==1.0.1
